add button start game
