+++ src/routes/Game/index.js
@@ -1,0 +1,16 @@
+const GamePage = () => {
+  // const handleClick = () => {
+  //   console.log('###: <Header />');
+  //   onClickButton && onClickButton('game');
+  // }
+
+  return (
+    <>
+      <div>This Game Page!!!</div>
+      {/* <button onClick={handleClick}>Home Page </button> */}
+      <button >Home Page </button>
+    </>
+  );
+};
+
+export default GamePage;
